+++ changelog.md
@@ -1,3 +1,7 @@
+## H2DE-3.2.7
+#### FIXED
+- flips and rotations
+
 ## H2DE-3.2.6
 #### FIXED
 - transparent texture with black default background